bug-2037388: Address review comments.
