+++ tecken/upload/client_otel.py
@@ -2,7 +2,6 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-import datetime
 from google.cloud import iam_credentials
 
 
@@ -21,7 +20,7 @@ class ClientOTelConfigProvider:
         self.log_level = log_level
         self.iam_client = iam_client or iam_credentials.IAMCredentialsClient()
 
-    def get(self) -> dict | None:
+    def get(self) -> dict:
         """Return the OTel config.
 
         Each call to this function generates a new access token for the OTel service
@@ -46,29 +45,7 @@ class ClientOTelConfigProvider:
 CONFIG: ClientOTelConfigProvider | None = None
 
 
-def init(
-    service_account: str,
-    gcp_project: str,
-    log_level: str,
-    iam_client=None,
-):
+def init(service_account: str, gcp_project: str, log_level: str):
     """Initialize the global OTel config provider."""
     global CONFIG
-    CONFIG = ClientOTelConfigProvider(
-        service_account, gcp_project, log_level, iam_client
-    )
-
-
-class MockIAMCredentialsClient:
-    """A mock implementation of IAMCredentialsClient for tests."""
-
-    def __init__(self, access_token):
-        self.access_token = access_token
-
-    def generate_access_token(
-        self, *, name: str, scope: list[str]
-    ) -> iam_credentials.GenerateAccessTokenResponse:
-        return iam_credentials.GenerateAccessTokenResponse(
-            access_token=self.access_token,
-            expire_time=datetime.datetime.now() + datetime.timedelta(seconds=3600),
-        )
+    CONFIG = ClientOTelConfigProvider(service_account, gcp_project, log_level)
